+++ apps/ui/src/components/Schema/Stats.tsx
