+++ apps/ui/src/components/Schema/StatsDetails.tsx
