+++ apps/ui/src/components/Schema/Type.tsx
