+++ apps/ui/src/components/Trace/Editor/Editor.tsx
