+++ apps/ui/src/components/Trace/Editor/json-viewer/JsonViewer.tsx
@@ -1,4 +1,4 @@
-import { IDS } from "../../testing";
+import { IDS } from "../../../../testing";
 import { JsonValue } from "./JsonValue";
 
 export function JsonViewer({ json }: { json: string }) {
