+++ apps/ui/src/components/Trace/Editor/json-viewer/index.ts
@@ -1,0 +1,2 @@
+export * from "./JsonValue";
+export * from "./JsonViewer";
