+++ apps/ui/src/components/Trace/Editor/query-viewer/Selection.tsx
