+++ apps/ui/src/components/Trace/Editor/query-viewer/Type.tsx
