+++ apps/ui/src/components/Trace/TraceStats/index.ts
@@ -1,0 +1,6 @@
+export * from "./SchemaTraces";
+export * from "./Span";
+export * from "./TraceHeader";
+export * from "./TraceStats";
+export * from "./TraceViewer";
+export * from "./utils";
