+++ apps/ui/src/components/Trace/index.ts
@@ -1,0 +1,3 @@
+export * from "./Editor";
+export * from "./TraceStats";
+export * from "./Trace";
