+++ apps/ui/src/components/index.ts
@@ -1,1 +1,4 @@
 export * from "./SideBar";
+export * from "./SchemaOverview/ListSchema";
+export * from "./SchemaOverview/NoSchema";
+export * from "./Schema";
